requirements.txt
anaconda-client==1.7.2
anaconda-navigator==1.9.12
appdirs==1.4.4
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828585465/work
asgiref==3.3.1
attrs==19.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.1
black==20.8b1
bleach==3.1.5
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
clyent==1.2.2
colorama==0.4.3
conda==4.9.0
conda-build==3.18.11
conda-package-handling==1.7.0
conda-verify==3.4.2
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.7
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
future==0.18.2
glob2==0.7
gunicorn==20.1.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446511143/work
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593447482397/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.1.0
jupyter-core==4.6.3
libarchive-c==2.9
lxml==4.5.2
markdown2==2.4.0
MarkupSafe==1.1.1
menuinst==1.4.16
mistune==0.8.4
mypy-extensions==0.4.3
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.7
notebook @ file:///C:/ci/notebook_1596840175976/work
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
pathspec==0.8.0
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1594298230227/work
pkginfo==1.5.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
psycopg2==2.8.6
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.6.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1592848611704/work
QtPy==1.9.0
regex==2020.10.23
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml==0.15.87
selenium==3.141.0
Send2Trash==1.5.0
sip==4.19.13
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1596810128862/work
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==1.2.0
xmltodict==0.12.0
zipp==3.1.0
